+++ src/provider/SemanticTokensProvider/tools/DAList2SemanticHighlight.ts
@@ -1,5 +1,7 @@
 import type { CAhkFunc } from '../../../AhkSymbol/CAhkFunc';
 import type { TAstRoot } from '../../../AhkSymbol/TAhkSymbolIn';
+import { hoverAVar } from '../../../tools/Built-in/1_built_in_var/A_Variables.tools';
+import { hoverBiVar } from '../../../tools/Built-in/1_built_in_var/BiVariables.tools';
 import { getDAListTop } from '../../../tools/DeepAnalysis/getDAList';
 import type { TSemanticTokensLeaf } from '../tools';
 
@@ -16,7 +18,7 @@ function DA2SemanticHighlight(fn: CAhkFunc): TSemanticTokensLeaf[] {
         }
     }
     for (const [k, { defRangeList, refRangeList }] of valMap) {
-        if (k === 'CLIPBOARD' || k === 'CLIPBOARDALL') continue;
+        if (hoverBiVar(k) !== undefined || hoverAVar(k) !== undefined) continue;
 
         for (const { range } of [...defRangeList, ...refRangeList]) {
             Tokens.push({