+++ src/provider/SemanticTokensProvider/tools/ModuleVarSemantic.ts
@@ -1,3 +1,5 @@
+import { hoverAVar } from '../../../tools/Built-in/1_built_in_var/A_Variables.tools';
+import { hoverBiVar } from '../../../tools/Built-in/1_built_in_var/BiVariables.tools';
 import type { TModuleVar } from '../../../tools/DeepAnalysis/getModuleVarMap';
 import type { TSemanticTokensLeaf } from '../tools';
 
@@ -5,7 +7,7 @@ export function ModuleVarSemantic(ModuleVar: TModuleVar): TSemanticTokensLeaf[] 
     const Tokens: TSemanticTokensLeaf[] = [];
     const { ModuleValMap } = ModuleVar;
     for (const [k, { defRangeList, refRangeList }] of ModuleValMap) {
-        if (k === 'CLIPBOARD' || k === 'CLIPBOARDALL') continue;
+        if (hoverBiVar(k) !== undefined || hoverAVar(k) !== undefined) continue;
 
         for (const { range } of [...defRangeList, ...refRangeList]) {
             Tokens.push({